Lift the quarantine review endpoints out of web_server
api/quarantine.py takes the /api/quarantine family plus
/api/review-queue/summary: list, stream, play, compare, audit, file
tags, one-click approve (with its sibling cleanup and retry-cancel),
recover-to-staging, delete, clear. it imports the two play-session
helpers from api/verification - the other half of the review queue -
so the back-import shim in web_server goes away again.

bodies byte-identical; decorators swapped, the two rebindable boot
globals became getters, download_tasks/tasks_lock come straight from
core.runtime_state, and the two post-process entry points are injected
as stable function refs. routemap identical on method+path, smoke green
on list/summary/entry-404, and the summary-endpoint source-pinning
tests repointed to the handler's new home (the documented fail-honestly
case). web_server: 36,999 -> 36,624 lines.
